parsers/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
antlr4-python3-runtime==4.13.2
    # via -r requirements.in
attrs==24.3.0
    # via
    #   flake8-bugbear
    #   flake8-mypy
bandit==1.8.2
    # via flake8-bandit
black==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via
    #   black
    #   pip-tools
cognitive-complexity==1.3.0
    # via flake8-cognitive-complexity
darglint==1.8.1
    # via -r requirements.in
dlint==0.16.0
    # via -r requirements.in
docutils==0.21.2
    # via restructuredtext-lint
flake8==7.1.1
    # via
    #   -r requirements.in
    #   dlint
    #   flake8-bandit
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-docstrings
    #   flake8-isort
    #   flake8-mutable
    #   flake8-mypy
    #   flake8-pep3101
    #   flake8-rst-docstrings
flake8-bandit==4.1.1
    # via -r requirements.in
flake8-bugbear==24.12.12
    # via -r requirements.in
flake8-builtins==2.5.0
    # via -r requirements.in
flake8-cognitive-complexity==0.1.0
    # via -r requirements.in
flake8-comprehensions==3.16.0
    # via -r requirements.in
flake8-docstrings==1.7.0
    # via -r requirements.in
flake8-isort==6.1.1
    # via -r requirements.in
flake8-mutable==1.2.0
    # via -r requirements.in
flake8-mypy==17.8.0
    # via -r requirements.in
flake8-pep3101==2.1.0
    # via -r requirements.in
flake8-rst-docstrings==0.3.0
    # via -r requirements.in
flake8-variables-names==0.0.6
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements.in
isort==5.13.2
    # via flake8-isort
jinja2==3.1.5
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.14.1
    # via flake8-mypy
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.2
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pbr==6.1.0
    # via stevedore
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.2.0
    # via flake8
pygments==2.19.1
    # via
    #   flake8-rst-docstrings
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyupgrade==3.19.1
    # via -r requirements.in
pyyaml==6.0.2
    # via bandit
restructuredtext-lint==1.4.0
    # via flake8-rst-docstrings
rich==13.9.4
    # via bandit
smmap==5.0.2
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
stevedore==5.4.0
    # via bandit
tokenize-rt==6.1.0
    # via pyupgrade
typing-extensions==4.12.2
    # via mypy
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
